requirements.txt
astroid==1.6.2
awsebcli==3.12.4
blessed==1.14.2
boto3==1.6.19
botocore==1.9.19
cement==2.8.2
colorama==0.3.7
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.3
django-cors-headers==2.4.0
django-debug-toolbar==1.8
django-extensions==2.0.7
django-heroku==0.3.1
django-storages==1.6.6
django-toolbelt==0.0.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
gunicorn==19.7.1
idna==2.7
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
model-mommy==1.6.0
multidict==4.4.0
olefile==0.46
pathspec==0.5.5
pep8==1.7.0
Pillow==4.3.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pudb==2018.1
Pygments==2.2.0
PyJWT==1.6.4
pylint==1.8.1
python-dateutil==2.6.1
python-decouple==3.1
pytz==2018.3
PyYAML==3.12
requests==2.9.1
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
tabulate==0.7.5
termcolor==1.1.0
urwid==2.0.1
vcrpy==1.11.1
wcwidth==0.1.7
websocket-client==0.47.0
whitenoise==4.1
wrapt==1.10.11
yarl==1.2.6
